Model - view separation
