+++ src/Views/PieceView.tsx
@@ -1,0 +1,7 @@
+
+export const PieceView = (props: { src: string }) => {
+
+    return <img className="piece-img" src={props.src} alt="" />;
+
+}
+
